+++ src/main/java/Main.java
@@ -1,13 +1,13 @@
 import cli.BootCLI;
 import cli.BootCLIInterface;
-import logInuseCase.InvalidInputException;
+import logInUseCase.InvalidInputException;
 
 import java.io.IOException;
 
 
 /**
  * The most important class of the project and at the same time is the shortest one. It is not supposed to
- * have any other calls other than to run the project.
+ * have any other calls other than to run the project. Tests are omitted for the same reason.
  */
 public class Main {
