+++ src/main/java/cli/LogInInterface.java
@@ -4,9 +4,10 @@ import java.io.IOException;
 import java.util.Scanner;
 import logInInterfaceAdapter.LogInController;
 import logInInterfaceAdapter.LogInPresenter;
-import logInuseCase.InvalidInputException;
+import logInUseCase.InvalidInputException;
 
 public interface LogInInterface {
 
-    void login(Scanner scanner, LogInController logInController, LogInPresenter logInPresenter) throws IOException, ClassNotFoundException, InvalidInputException;
+    void login(Scanner scanner, LogInController logInController, LogInPresenter logInPresenter)
+            throws IOException, ClassNotFoundException, InvalidInputException;
 }
