+++ src/main/java/cli/MainMenu.java
@@ -2,7 +2,7 @@ package cli;
 
 import logInInterfaceAdapter.LogInController;
 import logInInterfaceAdapter.LogInPresenter;
-import logInuseCase.InvalidInputException;
+import logInUseCase.InvalidInputException;
 
 import java.io.IOException;
 import java.util.Scanner;
@@ -10,8 +10,8 @@ import java.util.Scanner;
 public class MainMenu implements MainMenuInterface{
     @Override
     public void displayMainMenu(LogInPresenter logInPresenter) {
-        logInPresenter.outputMessage("Avaliable Actions" + "\n" + "1. Show all the Users on our forum" + "\n" +
-                "2. Log out" + "\n" + "Another features will be avaliable later..." + "\n");
+        logInPresenter.outputMessage("Available Actions" + "\n" + "1. Show all the Users on our forum" + "\n" +
+                "2. Log out" + "\n" + "Another features will be available later..." + "\n");
         logInPresenter.outputMessage("Please, enter the number of the option to proceed"+ "\n");
     }
 
@@ -34,6 +34,7 @@ public class MainMenu implements MainMenuInterface{
         } else if (choice == 2) {
             try{
                 logInController.userLogOut();
+                logInPresenter.outputMessage("Log out successful!"+ "\n");
                 WelcomeMenuInterface welcomeMenuInterface =new WelcomeMenu();
                 welcomeMenuInterface.choseLoginOrRegister(scanner, logInController, logInPresenter);
             }  catch (IOException  e) {
