+++ src/main/java/cli/RegisterInterface.java
@@ -4,10 +4,11 @@ import java.io.IOException;
 import java.util.Scanner;
 import logInInterfaceAdapter.LogInController;
 import logInInterfaceAdapter.LogInPresenter;
-import logInuseCase.InvalidInputException;
+import logInUseCase.InvalidInputException;
 
 public interface RegisterInterface {
 
-    void register(Scanner scanner, LogInController logInController, LogInPresenter logInPresenter) throws IOException, ClassNotFoundException, InvalidInputException;
+    void register(Scanner scanner, LogInController logInController, LogInPresenter logInPresenter)
+            throws IOException, ClassNotFoundException, InvalidInputException;
 
 }
