+++ src/main/java/cli/WelcomeMenu.java
@@ -2,7 +2,7 @@ package cli;
 
 import logInInterfaceAdapter.LogInController;
 import logInInterfaceAdapter.LogInPresenter;
-import logInuseCase.InvalidInputException;
+import logInUseCase.InvalidInputException;
 
 import java.io.IOException;
 import java.util.Scanner;
@@ -10,7 +10,7 @@ import java.util.Scanner;
 public class WelcomeMenu implements WelcomeMenuInterface{
     @Override
     public void displayWelcomeMenu(LogInPresenter logInPresenter) {
-        logInPresenter.outputMessage("\n" + "Avaliable Actions" + "\n" + "1. Register as new User" + "\n" +
+        logInPresenter.outputMessage("\n" + "Available Actions" + "\n" + "1. Register as new User" + "\n" +
                 "2. Log in (if you already have an account)" + "\n");
         logInPresenter.outputMessage("Please, enter the number of the option to proceed"+ "\n");
     }
