+++ src/main/java/cli/WelcomeMenuInterface.java
@@ -2,7 +2,7 @@ package cli;
 
 import logInInterfaceAdapter.LogInController;
 import logInInterfaceAdapter.LogInPresenter;
-import logInuseCase.InvalidInputException;
+import logInUseCase.InvalidInputException;
 
 import java.io.IOException;
 import java.util.Scanner;
@@ -11,6 +11,7 @@ public interface WelcomeMenuInterface {
 
     void displayWelcomeMenu(LogInPresenter logInPresenter);
 
-    void choseLoginOrRegister(Scanner scanner, LogInController logInController, LogInPresenter logInPresenter) throws IOException, ClassNotFoundException, InvalidInputException;
+    void choseLoginOrRegister(Scanner scanner, LogInController logInController, LogInPresenter logInPresenter)
+            throws IOException, ClassNotFoundException, InvalidInputException;
 
 }
