--- src/main/java/logInuseCase/InvalidInputException.java
@@ -1,9 +1,0 @@
-package logInuseCase;
-
-/**
- * this Exception might be customised later, but it is renamed to understand quickly what exactly
- * went wrong when debugging
- */
-public class InvalidInputException extends  Exception{
-
-}
