+++ src/test/java/logInUseCase/LogInInteractorTest.java
@@ -1,4 +1,4 @@
-package logInuseCase;
+package logInUseCase;
 
 import org.junit.jupiter.api.Test;
